Moved Emulator dir to the model dir
