+++ src/app/Model/Emulator/alu.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { RegProviderService } from '../../Model/reg-provider.service';
+import { RegProviderService } from '../reg-provider.service';
 import { BBusService } from './b-bus.service';
 
 @Injectable({
