+++ src/app/Model/Emulator/parser.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
-import { RegProviderService } from '../../Model/reg-provider.service';
-import { MicroTokenizerService } from '../micro-tokenizer.service';
-import { Token } from '../micro-tokenizer.service';
+import { RegProviderService } from '../reg-provider.service';
+import { MicroTokenizerService } from '../../Controller/micro-tokenizer.service';
+import { Token } from '../../Controller/micro-tokenizer.service';
 
 export interface Instruction {
   addr: Array<number>;
